+++ client/src/components/Users.jsx
@@ -7,34 +7,114 @@ import * as api from '../api';
 require('../styles/EntityTypeHeader.scss');
 require('../styles/Users.scss');
 
+class UserActionSelector extends Component {
+  constructor(props) {
+    super(props);
+    this.state = { action: '?' };
+    this.onChange = this.onChange.bind(this);
+  }
 
-function User({ email, username, admin }) {
+  onChange(event) {
+    const action = event.target.value;
+    const userId = this.props.userId;
+    this.setState({ action });
+    this.props.onChange(userId, action);
+  }
+
+  render() {
+    const { active, admin } = this.props;
+    return (
+      <select
+        className="UserActionSelector"
+        onChange={this.onChange}
+        value={this.state.action}
+      >
+        <option value="?">...</option>
+        <option disabled key="user-edit" value="edit">
+          Edit
+        </option>
+        <option disabled={active} key="user-delete" value="delete">
+          Delete
+        </option>
+        {!admin
+           ?
+             <option disabled={admin} key="user-promote" value="promote">
+               Enable admin privileges
+             </option>
+           :
+             <option disabled={(!admin || active)} key="user-demote" value="demote">
+               Remove admin privileges
+             </option>
+        }
+      </select>
+    );
+  }
+}
+
+UserActionSelector.propTypes = {
+  active: PropTypes.bool.isRequired,
+  admin: PropTypes.bool.isRequired,
+  onChange: PropTypes.func.isRequired,
+  userId: PropTypes.string.isRequired,
+};
+
+function User({ active, admin, email, onChange, userId, username }) {
   return (
     <tr>
-      <td>{username}</td>
+      <td>
+        {username}
+        {active
+          ? <span className="isMe"> (me)</span>
+          : <span />
+        }
+      </td>
       <td>{email}</td>
       <td>{admin ? 'Admin' : 'User'}</td>
+      <td>
+        <UserActionSelector
+          active={active}
+          admin={admin}
+          onChange={onChange}
+          userId={userId}
+        />
+      </td>
     </tr>
   );
 }
 
 User.propTypes = {
-  username: PropTypes.string.isRequired,
+  active: PropTypes.bool.isRequired,
+  admin: PropTypes.bool.isRequired,
   email: PropTypes.string.isRequired,
-  admin: PropTypes.bool,
+  onChange: PropTypes.func.isRequired,
+  userId: PropTypes.string.isRequired,
+  username: PropTypes.string.isRequired,
 };
 
 User.defaultProps = {
   admin: false,
 };
 
-function UserList({ users }) {
+function UserList({ activeUserId, onChange, users }) {
   return (
     <table>
       <tbody>
-        <tr><th>Name</th><th>Email</th><th>Role</th></tr>
-        {users.map(({ email, admin, username }) => (
-          <User key={username} email={email} admin={admin} username={username} />
+        <tr>
+          <th>Name</th>
+          <th>Email</th>
+          <th>Role</th>
+          <th>Actions</th>
+        </tr>
+        {users.map(({ admin, email, id, username }) => (
+          <User
+            active={id === activeUserId}
+            admin={admin}
+            email={email}
+            key={id}
+            onChange={onChange}
+            userId={id}
+            username={username}
+          />
         ))}
       </tbody>
     </table>
@@ -42,11 +122,12 @@ function UserList({ users }) {
 }
 
 UserList.propTypes = {
+  activeUserId: PropTypes.string.isRequired,
+  onChange: PropTypes.func.isRequired,
   users: PropTypes.array.isRequired,
 };
 
 class Users extends Component {
-
   constructor() {
     super();
     this.state = {
@@ -54,35 +135,55 @@ class Users extends Component {
       users: [],
     };
     this.getActionButtons = this.getActionButtons.bind(this);
+    this.getUsers = this.getUsers.bind(this);
     this.onInviteUser = this.onInviteUser.bind(this);
+    this.onUserActionChange = this.onUserActionChange.bind(this);
   }
 
   componentDidMount() {
     if (this.props.profile.admin) {
-      api.get('/api/admin/users')
-        .then(users => this.setState({ users }));
+      this.getUsers();
     }
   }
 
-  onInviteUser(emailAddress) {
+  onInviteUser(email) {
     this.setState({ isInviteModalVisible: false });
-    api.post('/api/admin/invites', {
-      email: emailAddress,
-    });
+    api.post('/api/admin/invites', { email });
+  }
+
+  onUserActionChange(userId, action) {
+    const url = `/api/admin/users/${userId}`;
+    if (action === 'delete') {
+      api.del(url).then(() => this.getUsers());
+    } else if (action === 'demote') {
+      api.patch(url, { admin: false }).then(() => this.getUsers());
+    } else if (action === 'promote') {
+      api.patch(url, { admin: true }).then(() => this.getUsers());
+    }
   }
 
   getActionButtons() {
-    const invite = {
-      buttonText: 'Invite user',
-      onClick: () => this.setState({ isInviteModalVisible: true }),
-    };
+    const buttons = [
+      {
+        buttonText: 'Manage invites',
+        onClick: null,
+      },
+      {
+        buttonText: 'Invite user',
+        onClick: () => this.setState({ isInviteModalVisible: true }),
+      },
+    ];
+    return buttons;
+  }
 
-    return [invite];
+  getUsers() {
+    api.get('/api/admin/users')
+      .then(users => this.setState({ users }));
   }
 
   render() {
     const actionButtons = this.getActionButtons();
-    const { admin } = this.props.profile;
+    const { admin, id } = this.props.profile;
     const saveStatus = '';
     const title = 'Members';
 
@@ -102,7 +203,11 @@ class Users extends Component {
           actionButtons={actionButtons}
         />
         <div className="UserList">
-          <UserList users={this.state.users} />
+          <UserList
+            activeUserId={id}
+            onChange={this.onUserActionChange}
+            users={this.state.users}
+          />
         </div>
         <InviteUser
           isOpen={this.state.isInviteModalVisible}
@@ -121,5 +226,6 @@ export default connect(state => ({
 Users.propTypes = {
   profile: PropTypes.shape({
     admin: PropTypes.bool,
+    id: PropTypes.string.isRequired,
   }).isRequired,
 };