+++ client/src/utils.js
@@ -1,0 +1,4 @@
+export default function isValidEmail(email) {
+  const regex = /^\w+([.-]?\w+)*@\w+([.-]?\w+)*(\.\w{2,9})+$/;
+  return regex.test(email);
+}